Fix operand ordering bug
Values were being pulled off of the stack and assigned to the
wrong variables, resulting in order dependent operations (subtract
 & divide) not working as expected. This was not caught by the
tests because they used examples that were not order dependent. For
instance, 3 3 -, which equals 0 regardless of whether you subtract
the first number from the second, or the second number from the
first. Division is similarly order dependent and the tests were
using an example that was not order dependent. This commit fixes
both the tests and the code.