Return stack when it contains > 1 number
As of this commit, the logic for the return value of the `result`
method is as follows:

- If there are no numbers on the stack, `nil` is returned
- If a single number is on the stack, that number is returned
- If > 1 numbers are on the stack, an array containing all of the
  values is returned

At this point in time, these return values feel like the most
semantically appropriate choice. I can see that perhaps it would
be best to just always return the stack and leave it to the client
to interpret the stack appropriately to craft user feedback. If
such an approach becomes necessary, I may just make the stack
publicly accessible via an `attr_reader`.