Add sleep to test method to reduce random failures
Because the tests are interacting with an actual, running instance
of the CLI, there are async issues sometimes when waiting for
output from the CLI. This sleeps the method which gets the
response in order to give the CLI time to write to stdout.

A better approach would be to attempt to repeatedly read the
response for either a certain amount of time, or a certain number
of attempts.